Move slug and title functions to shared utils file. Make sure the title is unique on conversion.
