+++ packages/edit-site/src/components/add-new-template/new-template-part.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { kebabCase } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { useState } from '@wordpress/element';
@@ -20,7 +15,11 @@ import { plus } from '@wordpress/icons';
 import { useHistory } from '../routes';
 import { store as editSiteStore } from '../../store';
 import CreateTemplatePartModal from '../create-template-part-modal';
-import { useExistingTemplateParts } from './utils';
+import {
+	useExistingTemplateParts,
+	getUniqueTemplatePartTitle,
+	getCleanTemplatePartSlug,
+} from '../../utils/template-part-create';
 
 export default function NewTemplatePart( {
 	postType,
@@ -42,39 +41,19 @@ export default function NewTemplatePart( {
 			return;
 		}
 
-		const uniqueTitle = () => {
-			const lowercaseTitle = title.toLowerCase();
-			const existingTitles = existingTemplateParts.map(
-				( templatePart ) => templatePart.title.rendered.toLowerCase()
-			);
-
-			if ( ! existingTitles.includes( lowercaseTitle ) ) {
-				return title;
-			}
-
-			let suffix = 2;
-			while (
-				existingTitles.includes( `${ lowercaseTitle } ${ suffix }` )
-			) {
-				suffix++;
-			}
-
-			return `${ title } ${ suffix }`;
-		};
-
 		try {
-			// Currently template parts only allow latin chars.
-			// Fallback slug will receive suffix by default.
-			const cleanSlug =
-				kebabCase( title ).replace( /[^\w-]+/g, '' ) ||
-				'wp-custom-part';
+			const cleanSlug = getCleanTemplatePartSlug( title );
+			const uniqueTitle = getUniqueTemplatePartTitle(
+				title,
+				existingTemplateParts
+			);
 
 			const templatePart = await saveEntityRecord(
 				'postType',
 				'wp_template_part',
 				{
 					slug: cleanSlug,
-					title: uniqueTitle(),
+					title: uniqueTitle,
 					content: '',
 					area,
 				},
