+++ packages/edit-site/src/components/add-new-template/utils.js
@@ -52,20 +52,6 @@ export const useExistingTemplates = () => {
 	);
 };
 
-export const useExistingTemplateParts = () => {
-	return useSelect(
-		( select ) =>
-			select( coreStore ).getEntityRecords(
-				'postType',
-				'wp_template_part',
-				{
-					per_page: -1,
-				}
-			),
-		[]
-	);
-};
-
 export const useDefaultTemplateTypes = () => {
 	return useSelect(
 		( select ) =>
